Criação de novos métodos

+++ projeto_POO-BD/src/main/java/com/api/projeto_poobd/models/MusicaModel.java
@@ -23,11 +23,11 @@ public class MusicaModel {
     @ManyToMany(mappedBy = "musicasNoRepertorio")
     private Set<RepertorioModel> repertorios = new HashSet<>();
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
